修改世界中、汇率换算闪退bug。

+++ destination/src/main/java/com/aibabel/traveladvisory/app/MyApplication.java
@@ -136,16 +136,19 @@ public class MyApplication extends Application {
             @Override
             public void onActivityCreated(Activity activity, Bundle savedInstanceState) {
                 activityLinkedList.add(activity);
+                canExit=false;
             }
 
             @Override
             public void onActivityStarted(Activity activity) {
                 Log.d(TAG, "onActivityStarted: " + activity.getLocalClassName());
                 stateCount++;
+                canExit=false;
             }
 
             @Override
             public void onActivityResumed(Activity activity) {
+                canExit=false;
             }
 
             @Override
@@ -169,21 +172,34 @@ public class MyApplication extends Application {
         });
     }
 
+    protected   static  volatile boolean canExit=true;
     /**
      * 退出所有app
      */
     public static void exit() {
-        for (Activity activity : activityLinkedList) {
-            activity.finish();
+        if(activityLinkedList!=null&& activityLinkedList.size()>0){
+            for (Activity activity : activityLinkedList) {
+                activity.finish();
+            }
         }
-        new Handler().postDelayed(new Runnable() {
+        canExit=true;
+        new Thread(new Runnable() {
             @Override
             public void run() {
-                android.os.Process.killProcess(android.os.Process.myPid());
-            }
-        },3000);
-    }
+                try {
+                    Thread.sleep(3000);
+                    Log.e("canExit===",String.valueOf(canExit));
+                    if ( canExit){
+                        android.os.Process.killProcess(android.os.Process.myPid());
+                    }
+                } catch (InterruptedException e) {
+                    e.printStackTrace();
+                }
 
+            }
+        }).start();
+
+    }
     public void initUmengConfig() {
         //初始化组件化基础库, 统计SDK/推送SDK/分享SDK都必须调用此初始化接口
         UMConfigure.init(this, "5b519f8cf29d981e1400006b", CommonUtils.getSN(), UMConfigure.DEVICE_TYPE_PHONE,
